1. Add test for authentication middleware 2. delete extra eslint comments 3. fix catch block format in product.js 4. remove /api/user/:userId/products route from public routes in the constants file 5. move the regestering validation middleware to the validation file 6. delete auth.js file from the middleware folder 7. rearrange routes in the auth.js file
