+++ src/config/passport.js
@@ -1,4 +1,3 @@
-/* eslint-disable prefer-template */
 const passport = require('passport');
 const GoogleStrategy = require('passport-google-oauth20').Strategy;
 const FacebookStrategy = require('passport-facebook').Strategy;
