+++ src/docs/Auth/OAuth2.yaml
@@ -3,7 +3,7 @@ paths:
     get:
       tags:
       - Authentication
-      summary: "Authenticates user using google 3rd party authentication service"
+      summary: "Authenticates users using google 3rd party authentication service"
       responses:
         "200":
           description: "Redirects verified user to the home page"
