+++ src/docs/Auth/facebookAuth.yaml
@@ -3,7 +3,7 @@ paths:
     get:
       tags:
       - Authentication
-      summary: "Authenticates user using Facebook 3rd party authentication service"
+      summary: "Authenticates users using Facebook 3rd party authentication service"
       responses:
         "200":
           description: "Redirects verified user to the profile page"
