+++ src/docs/Product/deleteProduct.yaml
@@ -3,6 +3,7 @@ paths:
     delete:
       tags:
         - Product
+      summary: "Deletes a product (Requires log in)"
       description: 'Deletes a product'
       parameters:
         - name: 'productId'
