+++ src/docs/User/deleteProfile.yaml
@@ -3,6 +3,7 @@ paths:
     delete:
       tags:
         - User
+      summary: Delete the profile of the logged in user (Requires log in)
       description: 'Delete the profile of the logged in user, including all products by this user'
       responses:
         '200':
