+++ src/docs/User/updateProfile.yaml
@@ -2,7 +2,7 @@ paths:
   /api/user/profile/update:
     patch:
       tags:
-        - Update User
+        - User
       requestBody:
         content:
           application/json:
@@ -34,11 +34,11 @@ paths:
                   "nationality": "Turkey",
                   "refugee": false
                   }'
-              'PHONENUMBER NOT INCORRECT FORMAT':
+              'INVALID PHONENUMBER FORMAT':
                 value: '{
                   "phoneNumber": "+60"
                   }'
-              'AGE IS NOT IN A VALID FORMAT':
+              'INVALID AGE FORMAT':
                 value: '{
                   "age": "225"
                   }'
@@ -53,7 +53,7 @@ paths:
                   "nationality": "String",
                   "refugee": false
                   }'
-      summary: 'Updates information of the user'
+      summary: Updates information of the user (Requires log in)
       description: 'Should update the info of the current user'
       produces:
         - 'application/json'
@@ -67,7 +67,7 @@ paths:
   /api/user/profile/update/password:
     patch:
       tags:
-        - Update User
+        - User
       requestBody:
         content:
           application/json:
@@ -111,7 +111,7 @@ paths:
                   "password": "String",
                   "passwordConfirmation": "String"
                   }'
-      summary: 'Updates password of the user'
+      summary: Updates password of the user (Requires log in)
       description: 'Should update the password of the current user'
       produces:
         - 'application/json'
