+++ src/routes/auth.js
@@ -1,18 +1,30 @@
 const express = require('express');
-const { registerMiddleware } = require('../middleware/auth');
 const authController = require('../controllers/auth');
 const { passport } = require('../config/passport');
 const {
   userLoginValidationRules,
   errorHandlingForValidation,
+  registerMiddleware,
 } = require('../middleware/validation');
 
 const router = express.Router();
+
+router.post('/register', registerMiddleware, authController.register);
+
+router.post(
+  '/login',
+  userLoginValidationRules,
+  errorHandlingForValidation,
+  authController.login
+);
+
+router.post('/logout', authController.logout);
 
 router.get(
   '/google',
   passport.authenticate('google', { scope: ['profile', 'email', 'openid'] })
 );
+
 router.get(
   '/google/callback',
   passport.authenticate('google', {
@@ -21,14 +33,6 @@ router.get(
   }),
   authController.saveUserToTokenAndCookie
 );
-router.post(
-  '/login',
-  userLoginValidationRules,
-  errorHandlingForValidation,
-  authController.login
-);
-router.post('/register', registerMiddleware, authController.register);
-router.post('/logout', authController.logout);
 
 router.get(
   '/facebook',
