+++ src/routes/product.js
@@ -5,7 +5,6 @@ const productController = require('../controllers/product');
 const router = express.Router();
 
 router.get('/', productController.getProducts);
-
 router.post('/', productMiddleware, productController.addNewProduct);
 router.delete('/:productId', productController.deleteProduct);
 router.patch('/:productId', productController.updateProduct);
